-[CHORE] Start to remove modules that are not required and move new elements to the global module until the final module is completed. Mining moved from Industry to App.
