+++ neocom-infinity-frontend/src/app/industry/pages/v1-mining-operations-page/V1MiningOperationsPage.ts
