+++ neocom-infinity-frontend/src/app/industry/pages/v1-mining-operations-page/v1-mining-operations-page.html
@@ -14,9 +14,9 @@
   </div> -->
   <div class="section-title">MINING - MINING OPERATIONS</div>
   <div class="row">
-    <v1-miningoperations-panel
+    <!-- <v1-miningoperations-panel
       class="panel"
       [container]="self"
-    ></v1-miningoperations-panel>
+    ></v1-miningoperations-panel> -->
   </div>
 </div>
