+++ neocom-infinity-frontend/src/app/industry/panels/v1-miningoperations-panel/v1-miningoperations-panel.component.html
